sender/requirements.txt
openpyxl==3.1.5
pandas==2.2.2
psycopg2==2.9.9
PyYAML==6.0.1